requirements.txt
streamlit
pandas
gspread
oauth2client
google-auth
google-auth-oauthlib
streamlit-components
reportlab
